+++ modules/@apostrophecms/asset/index.js
@@ -9,6 +9,8 @@ const express = require('express');
 const { stripIndent } = require('common-tags');
 const { merge: webpackMerge } = require('webpack-merge');
 const cuid = require('cuid');
+const chokidar = require('chokidar');
+const _ = require('lodash');
 const {
   checkModulesWebpackConfig,
   getWebpackExtensions,
@@ -29,7 +31,13 @@ module.exports = {
     // If this option is true and process.env.NODE_ENV is not `production`,
     // the browser will refresh when the Apostrophe application
     // restarts. A useful companion to `nodemon`.
-    refreshOnRestart: false
+    refreshOnRestart: false,
+    // If false no UI assets sources will be watched in development.
+    // This option has no effect in production (watch disabled).
+    watch: true,
+    // Miliseconds to wait between asset sources changes before
+    // performing a build.
+    watchDebounceMs: 1000
   },
 
   async init(self) {
@@ -48,26 +56,17 @@ module.exports = {
     self.extraBundles = fillExtraBundles(verifiedBundles);
     self.webpackExtensions = extensions;
     self.verifiedBundles = verifiedBundles;
+    self.buildWatcherEnable = process.env.APOS_ASSET_WATCH !== '0' && self.options.watch !== false;
+    self.buildWatcherDebounceMs = parseInt(self.options.watchDebounceMs || 1000, 10);
+    self.buildWatcher = null;
   },
   handlers (self) {
     return {
       'apostrophe:modulesRegistered': {
         async runUiBuildTask() {
-          if (
-            // Do not automatically build the UI if we're starting from a task
-            !self.apos.isTask() &&
-            // Or if we're in production
-            process.env.NODE_ENV !== 'production' &&
-            // Or if we've set an app option to skip the auto build
-            self.apos.options.autoBuild !== false
-          ) {
-
-            checkModulesWebpackConfig(self.apos.modules, self.apos.task.getReq().t);
-            // If starting up normally, run the build task, checking if we
-            // really need to update the apos build
-            await self.apos.task.invoke('@apostrophecms/asset:build', {
-              'check-apos-build': true
-            });
+          const ran = await self.autorunUiBuildTask();
+          if (ran) {
+            await self.watchUiAndRebuild();
           }
         },
         injectAssetsPlaceholders() {
@@ -79,6 +78,12 @@ module.exports = {
         async destroyUploadfs() {
           if (self.uploadfs && (self.uploadfs !== self.apos.uploadfs)) {
             await Promise.promisify(self.uploadfs.destroy)();
+          }
+        },
+        async destroyBuildWatcher() {
+          if (self.buildWatcher) {
+            await self.buildWatcher.close();
+            self.buildWatcher = null;
           }
         }
       }
@@ -867,6 +872,140 @@ module.exports = {
         return process.env.APOS_ASSET_CACHE ||
               path.join(self.apos.rootDir, 'data/temp/webpack-cache');
       },
+      // Run build task automatically when appropriate
+      async autorunUiBuildTask() {
+        if (
+        // Do not automatically build the UI if we're starting from a task
+          !self.apos.isTask() &&
+            // Or if we're in production
+            process.env.NODE_ENV !== 'production' &&
+            // Or if we've set an app option to skip the auto build
+            self.apos.options.autoBuild !== false
+        ) {
+
+          checkModulesWebpackConfig(self.apos.modules, self.apos.task.getReq().t);
+          // If starting up normally, run the build task, checking if we
+          // really need to update the apos build
+          await self.apos.task.invoke('@apostrophecms/asset:build', {
+            'check-apos-build': true
+          });
+          return true;
+        }
+        return false;
+      },
+      // Start watching assets from `modules/` and
+      // every symlinked package found in `node_modules/`.
+      // `rebuildCallback` is invoked with queue length argument
+      //  on actual build attempt only.
+      // It's there mainly for testing and debugging purposes.
+      async watchUiAndRebuild(rebuildCallback) {
+        if (!self.buildWatcherEnable || self.buildWatcher) {
+          return;
+        }
+        const rootDir = self.apos.rootDir;
+        // chokidar may invoke ready event multiple times,
+        // we want one "watch enabled" message.
+        let loggedOnce = false;
+        const logOnce = (...msg) => {
+          if (!loggedOnce) {
+            self.apos.util.log(...msg);
+            loggedOnce = true;
+          }
+        };
+        const error = self.apos.util.error;
+        const queue = [];
+        let queueLength = 0;
+        let queueRunning = false;
+
+        const debounceRebuild = _.debounce(chain, self.buildWatcherDebounceMs, {
+          leading: false,
+          trailing: true
+        });
+
+        const symLinkModules = await findSymlinks();
+        const watchDirs = [
+          './modules/**/ui/apos/**',
+          './modules/**/ui/src/**',
+          './modules/**/ui/public/**',
+          ...symLinkModules.reduce(
+            (prev, m) => [
+              ...prev,
+              `./node_modules/${m}/modules/**/ui/apos/**`,
+              `./node_modules/${m}/modules/**/ui/src/**`,
+              `./node_modules/${m}/modules/**/ui/public/**`
+            ],
+            []
+          )
+        ];
+        self.buildWatcher = chokidar.watch(watchDirs, {
+          cwd: rootDir,
+          ignoreInitial: true
+        });
+
+        self.buildWatcher
+          .on('add', debounceRebuild)
+          .on('change', debounceRebuild)
+          .on('unlink', debounceRebuild)
+          .on('addDir', debounceRebuild)
+          .on('unlinkDir', debounceRebuild)
+          .on('error', e => error(`Watcher error: ${e}`))
+          .on('ready', () => logOnce(
+            self.apos.task.getReq().t('apostrophe:assetBuildWatchStarted')
+          ));
+
+        async function rebuild() {
+          await self.autorunUiBuildTask();
+          self.restartId = self.apos.util.generateId();
+          if (typeof rebuildCallback === 'function') {
+            rebuildCallback(queueLength);
+          };
+        };
+
+        // Simple, capped, self-exhausting queue implementation.
+        function enqueue(fn) {
+          if (queueLength === 2) {
+            return;
+          }
+          queue.push(fn);
+          queueLength++;
+        };
+        async function dequeue() {
+          if (!queueLength) {
+            queueRunning = false;
+            return;
+          }
+          queueRunning = true;
+          await queue.pop()();
+          queueLength--;
+          await dequeue();
+        }
+        async function chain(f) {
+          enqueue(rebuild);
+          if (!queueRunning) {
+            await dequeue();
+          }
+        }
+
+        // Find all symlinks in node modules.
+        // This would find both `module-name` and `@company/module-name`
+        // package symlinks
+        async function findSymlinks(sub = '') {
+          let result = [];
+          const handle = await fs.promises.opendir(path.join(rootDir, 'node_modules', sub));
+          let mod = await handle.read();
+          while (mod) {
+            if (mod.isSymbolicLink()) {
+              result.push(sub + mod.name);
+            } else if (!sub && mod.name.startsWith('@')) {
+              const dres = await findSymlinks(`${mod.name}/`);
+              result = [ ...result, ...dres ];
+            }
+            mod = await handle.read();
+          }
+          await handle.close();
+          return result;
+        }
+      },
       // An implementation method that you should not need to call.
       // Sets a predetermined configuration for the frontend builds.
       // If you are trying to enable IE11 support for ui/src, use the
@@ -1013,9 +1152,17 @@ module.exports = {
           // Long polling: keep the logs quiet by responding slowly, except the
           // first time. If we restart, the request will fail immediately,
           // and the client will know to try again with `fast`. The client also
-          // uses `fast` the first time
-          if (!req.query.fast) {
-            await Promise.delay(30000);
+          // uses `fast` the first time.
+          if (req.query.fast) {
+            return self.restartId;
+          }
+          // Long polling will be interrupted if restartId changes.
+          let delay = 30000;
+          const step = 300;
+          const oldRestartId = self.restartId;
+          while (delay > 0 && oldRestartId === self.restartId) {
+            delay -= step;
+            await Promise.delay(step);
           }
           return self.restartId;
         }