+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Adds
 
 * Wepback cache for build performance in development and production mode.
+* Add a watcher in development mode to rebuild the assets on changes in the same process.
 * Add asset task `apostrophecms/asset:clear-cache` for force clearing the webpack build cache.
 
 ### Fixes